Cosine function

This function computes the approximation to the cosine function.

fn_cos(x, k)

Arguments

x

A numeric() number of length N.

k

An integer() that is greater than or equal to 1.

Value

An integer() or numeric() number that is the cosine of x.

Details

This function computes the cosine of a number using the truncated series expansion

Examples

## Generate some input data
x <- `2`
#> Error in eval(expr, envir, enclos): object '2' not found
k <- "10"

## Perform the cosine calculation
fn_cos(x, k)
#> Called from: eval(expr, p)
#> debug at /Users/yuhanxiao/RStudio/statprogramming/functions/R/fn_cos.R#21: if (k != as.integer(k)) {
#>     stop("Error: k should be an integer.")
#> } else if (k < 1) {
#>     stop("Error: k should be equal to or greater than 1")
#> }
#> debug at /Users/yuhanxiao/RStudio/statprogramming/functions/R/fn_cos.R#21: if (k < 1) {
#>     stop("Error: k should be equal to or greater than 1")
#> }
#> Error in eval(expr, envir, enclos): object 'x' not found